pyproject.toml
[project]
name = "omen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.0",
    "mypy>=1.15.0",
    "prometheus-client>=0.21.1",
    "pyyaml>=6.0.2",
]
